Changed sensors ontology and samples.
Primary keys now have to match "one or more times".  I.e. two objects
are indexed into the same location if they have the same selection of
primary key properties and the values for those properties are the same.
This should be extended into a flexible system with required keys and
these type of keys, so developers can mix and match for each class.