Many things; adding the new TestSensorReasoner.
And all that stuff surrounding that new class.  Also updated ontologies
to add reasoner result classes.  We still need to bring the two
different ontologies together with one shared vocabulary.

The Client interface has now a basic abstract implementation to aid
debugging (setName method).

DatabaseDataManager now uses exceptions better and shutdown also closes
the database connection.  If we lose the XMPP connection, we also
shutdown the program, although that should be revised to reconnect in
intervals or so.  Not sure how do that properly with regard to
synchronization though.